Vectorize SIMD.FloatPointsToF24Dot8Points

+++ SharpBlaze/LineArrayX32Y16.Inlines.cs
@@ -95,8 +95,8 @@ public unsafe partial struct LineArrayX32Y16
         {
             // Most common.
             current->Y0Y1[count] = y0y1;
-            current->X0[count] = (uint) x0;
-            current->X1[count] = (uint) x1;
+            current->X0[count] = x0;
+            current->X1[count] = x1;
 
             mCount = count + 1;
         }
@@ -105,8 +105,8 @@ public unsafe partial struct LineArrayX32Y16
             LineArrayX32Y16Block* b = memory.FrameNewX32Y16Block(current);
 
             b->Y0Y1[0] = y0y1;
-            b->X0[0] = (uint) x0;
-            b->X1[0] = (uint) x1;
+            b->X0[0] = x0;
+            b->X1[0] = x1;
 
             // Set count to 1 for segment being added.
             mCount = 1;
